feat: adjust filter bar based on design updates
