+++ src/components/EarnFilterBar/EarnFilterBarSkeleton.tsx
@@ -4,24 +4,35 @@ import { BaseSkeleton, EarnFilterBarContainer } from './EarnFilterBar.styles';
 export const EarnFilterBarSkeleton = () => {
   return (
     <EarnFilterBarContainer sx={(theme) => ({ gap: theme.spacing(1) })}>
-      <Stack direction="row" sx={(theme) => ({ gap: theme.spacing(1) })}>
-        {Array.from({ length: 2 }).map((_, index) => (
-          <BaseSkeleton key={index} variant="rounded" width={104} height={32} />
-        ))}
-      </Stack>
       <Stack
         direction="row"
-        sx={[
-          {
-            alignItems: 'center',
-          },
-          (theme) => ({ gap: theme.spacing(2) }),
-        ]}
+        sx={{
+          display: 'flex',
+          justifyContent: 'space-between',
+          alignItems: 'center',
+        }}
       >
-        {Array.from({ length: 5 }).map((_, index) => (
-          <BaseSkeleton key={index} variant="rounded" width={56} height={32} />
-        ))}
-        <BaseSkeleton variant="circular" width={40} height={40} />
+        <Stack direction="row" sx={(theme) => ({ gap: theme.spacing(1) })}>
+          {Array.from({ length: 2 }).map((_, index) => (
+            <BaseSkeleton
+              key={index}
+              variant="rounded"
+              width={104}
+              height={32}
+            />
+          ))}
+        </Stack>
+        <Stack
+          direction="row"
+          sx={(theme) => ({
+            gap: theme.spacing(2),
+            display: 'flex',
+            alignItems: 'center',
+          })}
+        >
+          <BaseSkeleton variant="rounded" width={56} height={32} />
+          <BaseSkeleton variant="circular" width={40} height={40} />
+        </Stack>
       </Stack>
     </EarnFilterBarContainer>
   );
