+++ src/components/EarnFilterBar/utils.ts
@@ -1,0 +1,24 @@
+import { formatSliderValue } from '@/components/core/form/Select/utils';
+import { toFixedFractionDigits } from '@/utils/formatNumbers';
+
+export const countBadge = (count: number): string | undefined =>
+  count > 0 ? count.toString() : undefined;
+
+export const valueBadge = (
+  usedMin: number,
+  usedMax: number,
+  rangeMin: number,
+  rangeMax: number,
+  pendingValue: number[],
+): string | undefined => {
+  if (
+    isNaN(usedMin) ||
+    isNaN(usedMax) ||
+    (usedMin === rangeMin && usedMax === rangeMax)
+  ) {
+    return undefined;
+  }
+  return formatSliderValue(
+    pendingValue.map((v) => toFixedFractionDigits(v, 0, 2)),
+  );
+};
