+++ src/i18n/translations/en/translation.json
@@ -432,6 +432,7 @@
       "selected": "{{count}} selected",
       "search": "Search {{filterBy}}...",
       "filterAndSort": "Filter and sort",
+      "filterSort": "Filters & Sort",
       "clearAll": "Clear all",
       "chain": "Chain",
       "protocol": "Protocol",
